+ geofirestore gradle, dashboard implementations

+++ app/src/main/java/com/devssocial/localodge/models/Location.kt
@@ -1,6 +1,6 @@
 package com.devssocial.localodge.models
 
 data class Location(
-    var lat: Float = 0f,
-    var lng: Float = 0f
+    var lat: Double = 0.0,
+    var lng: Double = 0.0
 )
